Simplify: extract pure cast helpers to cast-utils (#19 slice 1)
Move the stateless cast helpers — isChromecastSupported / the supported-mime
list, showCastError (throttled alert), and normalizeVolumeTo/FromCast — out of
the 824-line useCast.shared.ts monolith into lib/cast/cast-utils.ts. No React
state lives in these, so it's pure code motion: useCast.shared.ts imports them
instead. This carves the stateless seam ahead of the eventual
useCastDevices/useCastSession split (the entangled connect↔devices and
device-found↔connectedDevice closures are deferred until they can be exercised
on a real Chromecast — emulators can't validate the cast crash-path logic).

Also re-export TranscodeStatus from the useCast.ts shim so lib/cast/index.ts
resolves it (clears a pre-existing tsc error).

Validated: app tsc shows zero new errors and one fewer than baseline.
