+++ packages/app/lib/cast/useCast.ts
@@ -1,3 +1,3 @@
 export { useCast } from './useCast.shared'
-export type { CastDevice, CastPlaybackState, UseCastOptions, UseCastReturn } from './useCast.shared'
+export type { CastDevice, CastPlaybackState, TranscodeStatus, UseCastOptions, UseCastReturn } from './useCast.shared'
 export { useCast as default } from './useCast.shared'
